Create & modify README doc, HTML, CSS, & JS files
Created HTML, CSS and JS files for the necessary code to build the
calculator web app. Created a README file as well as adding a
description of the project. Added basic math functions to the JS file.
Added basic HTML in it's relative file to give the app a skeleton.