Modify button functionality & add more functions
Revised functions that were storing numbers, displaying numbers and
calculating equation answers. Added new functions for storing the
operators and actually displaying the answer on the calculator (this
wasn't happening before). Reintroduced function that cleared the
display for cleaner code with the button functions. Reintroduced the
splitting of the single button function for all buttons into 4 separate
functions representing numbers, operators, the equal and the clear
button. Modified DOM targeting at the top of JS to reflect the split.