Modify display styling & functionality
Removed HTML index form styling on display as it was too restrictive in
the display functionality on the JS side. Edited the target of the
display styling in CSS to reflect this. Prior to this the index form
display would disappear as soon as a number was entered. This issue
has been resolved. Fixed some of the functionality of the equals button
but not quite resolved yet as display clears without showing the answer
to the equation entered it.