Modify operator variable to new name & functions
Needed to change the operator value to currentOperator as it was
conflicting with several functions and other operations that tried to
use the "built in" operator symbol and the "operator" variable that I
previously declared and JS couldn't tell which to use when. Modified
storeOperator and answer functions to reflect these changes and correct
operator and equals button functionality.