Modify calculation function & operator buttons
Resolved issue of calculator not displaying the answer after the
equals sign was clicked. This was due to the answer() function calling
the function that stored the operator during runtime when it needed to
call the function that stored the number. The operator buttons also
were causing this issue as the parameter for the storeOperator()
function targeted the buttons by dataset. Changed this to target the
buttons by id, which resolved any more issues that the operator buttons
had.